backend/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic
